Refactored events logic to be clearer in config
added event triggering from within the module for potentially having the VA panel subscribe to the events it fires. This would reduce the need for it be edited and changed when integrated into the CV solution. They would simply subscribe to when the module thinks it should be showing or hiding the panel and respond accordingly.

This means we avoid mixing code specific to their panel with the other parts around the button status etc.